File drop area css harmonization and improvement
